Removes files with same name, type and size
Good for removing duplicate files. Does not work if duplicate files have different names or file types, for example if you copied and renamed a video ten times then it would not work. Also will delete files if they have the exact same name, file size, file type and possibly date so be careful because this isn't actually that rare.